safari-600.7-branch/LayoutTests/editing/pasteboard/contenteditable-pre.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This test copies and pastes content inside pre that is an editing host. WebKit should not clone pre.
To manually test, copy and paste "hello" and then paste it into the boxes below.
WebKit should not clone pre (nest pre) and the pasted content should not have nested borders.</p>
<style> #container pre, #container div {border: solid 2px blue;} </style>
<div id="container">
<pre contenteditable>hello</pre>
<div contenteditable></div>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description(document.getElementById('description').textContent);

var container = document.querySelector('#container pre');
container.focus();
document.execCommand('selectAll', false, null);
document.execCommand('copy', false, null);
document.execCommand('paste', false, null);

Markup.dump(container, "Pasting into pre; the pre should not be listed below");

var container = document.querySelector('#container div');
container.focus();
document.execCommand('paste', false, null);

Markup.dump(container, "Pasting into div; the pasted content should be in pre");

</script>
</body>
</html>
