third_party/blink/web_tests/editing/pasteboard/contenteditable-pre-expected.txt

This test copies and pastes content inside pre that is an editing host. WebKit should not clone pre.
To manually test, copy and paste "hello" and then paste it into the boxes below.
WebKit should not clone pre (nest pre) and the pasted content should not have nested borders.

Pasting into pre; the pre should not be listed below:
| "hello<#selection-caret>"

Pasting into div; the pasted content should be in pre:
| <pre>
|   "hello<#selection-caret>"
